+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/DOTS/EntityInstantiator.cs
@@ -38,7 +38,7 @@ namespace GLTFast {
         EntityArchetype m_NodeArchetype;
         EntityArchetype m_SceneArchetype;
 
-        Parent m_SceneParent;
+        List<Entity> m_Entities;
 
         public EntityInstantiator(
             IGltfReadable gltf,
@@ -59,6 +59,7 @@ namespace GLTFast {
             uint[] nodeIndices
         ) {
             Profiler.BeginSample("BeginScene");
+            m_Entities = new List<Entity>();
             m_Nodes = new Dictionary<uint, Entity>();
             m_EntityManager = World.DefaultGameObjectInjectionWorld.EntityManager;
             m_NodeArchetype = m_EntityManager.CreateArchetype(
@@ -69,24 +70,24 @@ namespace GLTFast {
             );
             m_SceneArchetype = m_EntityManager.CreateArchetype(
                 typeof(Disabled),
-                typeof(LocalTransform)
+                typeof(LocalTransform),
+                typeof(LocalToWorld)
             );
 
-            if (m_Settings.SceneObjectCreation == SceneObjectCreation.Never
-                || m_Settings.SceneObjectCreation == SceneObjectCreation.WhenMultipleRootNodes && nodeIndices.Length == 1) {
-                m_SceneParent.Value = m_Parent;
-            }
-            else {
-                var sceneEntity = m_EntityManager.CreateEntity(m_Parent==Entity.Null ? m_SceneArchetype : m_NodeArchetype);
-                m_EntityManager.SetComponentData(sceneEntity,LocalTransform.Identity);
-                m_EntityManager.SetComponentData(sceneEntity, new LocalToWorld{Value = float4x4.identity});
+            var dedicatedSceneEntity = m_Settings.SceneObjectCreation == SceneObjectCreation.Always
+                || m_Settings.SceneObjectCreation == SceneObjectCreation.WhenMultipleRootNodes && nodeIndices.Length > 1;
+
+            if (dedicatedSceneEntity) {
+                var sceneEntity = m_EntityManager.CreateEntity(m_Parent == Entity.Null ? m_SceneArchetype : m_NodeArchetype);
+                m_EntityManager.SetComponentData(sceneEntity, LocalTransform.Identity);
+                m_EntityManager.SetComponentData(sceneEntity, new LocalToWorld { Value = float4x4.identity });
 #if UNITY_EDITOR
                 m_EntityManager.SetName(sceneEntity, name ?? "Scene");
 #endif
                 if (m_Parent != Entity.Null) {
                     m_EntityManager.SetComponentData(sceneEntity, new Parent { Value = m_Parent });
                 }
-                m_SceneParent.Value = sceneEntity;
+                m_Entities.Add(sceneEntity);
             }
             Profiler.EndSample();
         }
@@ -107,9 +108,24 @@ namespace GLTFast {
             Vector3 position,
             Quaternion rotation,
             Vector3 scale
-        ) {
+        )
+        {
+            var parent = new Parent { Value = parentIndex.HasValue ? m_Nodes[parentIndex.Value] : m_Parent };
+            var node = CreateNodeInternal(parent, position, rotation, scale);
+            m_Nodes[nodeIndex] = node;
+        }
+
+        Entity CreateNodeInternal(
+            Parent parent,
+            float3 position,
+            quaternion rotation,
+            float3 scale
+        )
+        {
             Profiler.BeginSample("CreateNode");
-            var node = m_EntityManager.CreateEntity(m_NodeArchetype);
+            var validParent = parent.Value != Entity.Null;
+            var node = m_EntityManager.CreateEntity(validParent ? m_NodeArchetype : m_SceneArchetype);
+            m_Entities.Add(node);
             var isUniformScale = IsUniform(scale);
             m_EntityManager.SetComponentData(
                 node,
@@ -128,14 +144,13 @@ namespace GLTFast {
                     new PostTransformMatrix { Value = float4x4.Scale(scale) }
                     );
             }
-            m_Nodes[nodeIndex] = node;
-            m_EntityManager.SetComponentData(
-                node,
-                parentIndex.HasValue
-                    ? new Parent { Value = m_Nodes[parentIndex.Value] }
-                    : m_SceneParent
-                );
+
+            if (validParent)
+            {
+                m_EntityManager.SetComponentData(node, parent);
+            }
             Profiler.EndSample();
+            return node;
         }
 
         public void SetNodeName(uint nodeIndex, string name) {
@@ -158,15 +173,6 @@ namespace GLTFast {
                 return;
             }
             Profiler.BeginSample("AddPrimitive");
-            Entity node;
-            if(meshNumeration==0) {
-                // Use Node GameObject for first Primitive
-                node = m_Nodes[nodeIndex];
-            } else {
-                node = m_EntityManager.CreateEntity(m_NodeArchetype);
-                m_EntityManager.SetComponentData(node,LocalTransform.Identity);
-                m_EntityManager.SetComponentData(node, new Parent { Value = m_Nodes[nodeIndex] });
-            }
 
             var materials = new Material[meshResult.materialIndices.Length];
             for (var index = 0; index < meshResult.materialIndices.Length; index++)
@@ -189,6 +195,23 @@ namespace GLTFast {
 
             for (ushort index = 0; index < meshResult.materialIndices.Length; index++)
             {
+                Entity node;
+                if (meshNumeration == 0 && index == 0) {
+                    // Use node entity for first sub-mesh of first mesh result.
+                    node = m_Nodes[nodeIndex];
+                } else {
+                    node = CreateNodeInternal(
+                        new Parent { Value = m_Nodes[nodeIndex] },
+                        float3.zero,
+                        quaternion.identity,
+                        new float3(1f)
+                        );
+#if UNITY_EDITOR
+                    m_EntityManager.SetName(node, $"Entity-{meshNumeration}-submesh-{index}");
+#endif
+                    m_EntityManager.SetEnabled(node, true);
+                }
+
                 RenderMeshUtility.AddComponents(
                     node,
                     m_EntityManager,
@@ -197,15 +220,20 @@ namespace GLTFast {
                     MaterialMeshInfo.FromRenderMeshArrayIndices(
                         index,
                         0,
-#if UNITY_ENTITIES_1_2_OR_NEWER
-                        index
-#else
-                        (sbyte)index
+#if !UNITY_ENTITIES_1_2_OR_NEWER
+                        (sbyte)
 #endif
+                        index
                         )
                     );
 
-                m_EntityManager.SetComponentData(node, new RenderBounds {Value = meshResult.mesh.bounds.ToAABB()} );
+                // Refine RenderBounds
+                // RenderMeshUtility.AddComponents above already sets RenderBounds, but for the entire mesh.
+                if (meshResult.mesh.subMeshCount > 1)
+                {
+                    var subMesh = meshResult.mesh.GetSubMesh(index);
+                    m_EntityManager.SetComponentData(node, new RenderBounds { Value = subMesh.bounds.ToAABB() });
+                }
             }
 
             Profiler.EndSample();
@@ -244,56 +272,72 @@ namespace GLTFast {
                 LightProbeUsage = LightProbeUsage.Off,
             };
 
+            var prototype = m_EntityManager.CreateEntity(m_NodeArchetype);
+
+            if (scales.HasValue)
+            {
+                var scale = scales.Value[0];
+                m_EntityManager.AddComponent<PostTransformMatrix>(prototype);
+                m_EntityManager.SetComponentData(
+                    prototype,
+                    new PostTransformMatrix { Value = float4x4.Scale(scale) }
+                    );
+            }
+
+            var transform = new LocalTransform
+            {
+                Position = positions?[0] ?? Vector3.zero,
+                Rotation = rotations?[0] ?? Quaternion.identity,
+                Scale = 1
+            };
+            m_EntityManager.AddComponent<LocalTransform>(prototype);
+            m_EntityManager.SetComponentData(prototype, transform);
+
+            m_EntityManager.AddComponent<Parent>(prototype);
+            m_EntityManager.SetComponentData(prototype, new Parent { Value = m_Nodes[nodeIndex] });
+
             var renderMeshArray = new RenderMeshArray(materials, new[] { meshResult.mesh });
+            RenderMeshUtility.AddComponents(
+                prototype,
+                m_EntityManager,
+                renderMeshDescription,
+                renderMeshArray,
+                MaterialMeshInfo.FromRenderMeshArrayIndices(0, 0)
+            );
+
             for (ushort index = 0; index < meshResult.materialIndices.Length; index++)
             {
-                var prototype = m_EntityManager.CreateEntity(m_NodeArchetype);
-                m_EntityManager.SetEnabled(prototype, true);
-
                 for (var i = 0; i < instanceCount; i++) {
-                    var instance = i>0 ? m_EntityManager.Instantiate(prototype) : prototype;
 
-                    var transform = new LocalTransform
-                    {
-                        Position = positions?[i] ?? Vector3.zero,
-                        Rotation = rotations?[i] ?? Quaternion.identity,
-                        Scale = 1
-                    };
+                    var instance = index == 0 && i == 0 ? prototype : m_EntityManager.Instantiate(prototype);
+                    m_Entities.Add(instance);
+
+                    if (positions.HasValue)
+                        transform.Position = positions.Value[i];
+                    if (rotations.HasValue)
+                        transform.Rotation = rotations.Value[i];
                     if (scales.HasValue)
                     {
-                        var scale = scales.Value[i];
-                        var isUniformScale = IsUniform(scale);
-                        if (!isUniformScale)
-                        {
-                            m_EntityManager.AddComponent<PostTransformMatrix>(instance);
-                            m_EntityManager.SetComponentData(instance,new PostTransformMatrix {Value = float4x4.Scale(scale)});
-                        }
-                        else
-                        {
-                            transform.Scale = scale.x;
-                        }
+                        m_EntityManager.SetComponentData(
+                            instance,
+                            new PostTransformMatrix { Value = float4x4.Scale(scales.Value[i]) }
+                            );
                     }
-
-                    m_EntityManager.SetComponentData(instance,transform);
-                    m_EntityManager.SetComponentData(instance, new Parent { Value = m_Nodes[nodeIndex] });
+                    m_EntityManager.SetComponentData(instance, transform);
 
                     RenderMeshUtility.AddComponents(
                         instance,
                         m_EntityManager,
                         renderMeshDescription,
                         renderMeshArray,
-                        MaterialMeshInfo.FromRenderMeshArrayIndices(
-                            index,
-                            0,
-#if UNITY_ENTITIES_1_2_OR_NEWER
-                            index
-#else
-                            (sbyte)index
+                        MaterialMeshInfo.FromRenderMeshArrayIndices(index, 0,
+#if !UNITY_ENTITIES_1_2_OR_NEWER
+                            (sbyte)
 #endif
+                            index
                         )
                     );
                 }
-
             }
             Profiler.EndSample();
         }
@@ -321,10 +365,19 @@ namespace GLTFast {
         /// <inheritdoc />
         public virtual void EndScene(uint[] rootNodeIndices) {
             Profiler.BeginSample("EndScene");
-            m_EntityManager.SetEnabled(m_SceneParent.Value, true);
-            foreach (var entity in m_Nodes.Values) {
-                m_EntityManager.SetEnabled(entity, true);
+
+            if (m_Entities.Count > 0)
+            {
+                var mainEntity = m_Entities[0];
+                var entityGroup = m_EntityManager.AddBuffer<LinkedEntityGroup>(mainEntity);
+                entityGroup.Capacity = m_Entities.Count;
+                foreach (var entity in m_Entities)
+                {
+                    entityGroup.Add(new LinkedEntityGroup { Value = entity });
+                }
+                m_EntityManager.SetEnabled(mainEntity, true);
             }
+
             Profiler.EndSample();
         }